+++ R/list-modify.R
@@ -1,7 +1,7 @@
 #' Modify a list
 #'
-#' `modify_list()` recursively modifies a list, matching elements either by
-#' name or position. `update_list()` is a helper optimised for interactively
+#' `list_modify()` recursively modifies a list, matching elements either by
+#' name or position. `list_update()` is a helper optimised for interactively
 #' modifying an existing list: you can use a formula (a quosure) to replace
 #' using existing values.
 #'
@@ -11,10 +11,28 @@
 #' @param y Replacement list
 #' @export
 #' @examples
-#' x <- list(x = 1:10, y = 4)
-#' update_list(x, z = 10)
-#' update_list(x, z = ~ x + y)
-modify_list <- function(x, y) {
+#' x <- list(x = 1:10, y = 4, z = list(a = 1, b = 2))
+#' str(x)
+#'
+#' # Update values
+#' str(list_modify(x, list(a = 1)))
+#' # Replace values
+#' str(list_modify(x, list(z = 5)))
+#' str(list_modify(x, list(z = list(a = 1:5))))
+#' # Remove values
+#' str(list_modify(x, list(z = NULL)))
+#'
+#' # list_update is useful for interactive tweaking, because it
+#' # uses ... instead of a separate list
+#' str(list_update(x, a = 1))
+#' str(list_update(x, z = 5))
+#' str(list_update(x, z = list(a = 1:5)))
+#' str(list_update(x, z = NULL))
+#'
+#' # In list_update() you can also use formulas to compute new values
+#' list_update(x, z1 = ~ z[1])
+#' list_update(x, z = ~ x + y)
+list_modify <- function(x, y) {
   stopifnot(is.list(x), is.list(y))
 
   if (length(x) == 0) {
@@ -29,7 +47,7 @@ modify_list <- function(x, y) {
   if (is.null(x_names) && is.null(y_names)) {
     for (i in rev(seq_along(y))) {
       if (i <= length(x) && is_list(x[[i]]) && is_list(y[[i]])) {
-        x[[i]] <- modify_list(x[[i]], y[[i]])
+        x[[i]] <- list_modify(x[[i]], y[[i]])
       } else {
         x[[i]] <- y[[i]]
       }
@@ -37,7 +55,7 @@ modify_list <- function(x, y) {
   } else if (!is.null(x_names) && !is.null(y_names)) {
     for (nm in y_names) {
       if (has_name(x, nm) && is_list(x[[nm]]) && is_list(y[[nm]])) {
-        x[[nm]] <- modify_list(x[[nm]], y[[nm]])
+        x[[nm]] <- list_modify(x[[nm]], y[[nm]])
       } else {
         x[[nm]] <- y[[nm]]
       }
@@ -51,12 +69,18 @@ modify_list <- function(x, y) {
 
 
 #' @export
-#' @rdname modify_list
-update_list <- function(`_x`, ...) {
+#' @rdname  list_modify
+list_update <- function(`_x`, ...) {
   y <- list(...)
 
   needs_eval <- map_lgl(y, is_quosure)
   y[needs_eval] <- tidy_eval(y[needs_eval], `_x`)
 
-  modify_list(`_x`, y)
+  list_modify(`_x`, y)
 }
+
+#' @export
+#' @usage NULL
+#' @rdname list_modify
+update_list <- list_update
+